moved CmtcCommand into core
