+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/GotoExercise.scala
@@ -10,8 +10,9 @@ import com.lunatech.cmt.{
   toExecuteCommandErrorMessage
 }
 import com.lunatech.cmt.Helpers.{exerciseFileHasBeenModified, getFilesToCopyAndDelete, pullTestCode}
+import com.lunatech.cmt.Domain.StudentifiedRepo
 import com.lunatech.cmt.client.Configuration
-import com.lunatech.cmt.client.Domain.{ExerciseId, ForceMoveToExercise, StudentifiedRepo}
+import com.lunatech.cmt.client.Domain.{ExerciseId, ForceMoveToExercise}
 import com.lunatech.cmt.client.command.getCurrentExerciseId
 import com.lunatech.cmt.core.validation.Validatable
 import sbt.io.syntax.*
