+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/NextExercise.scala
@@ -10,8 +10,9 @@ import com.lunatech.cmt.{
   toExecuteCommandErrorMessage
 }
 import com.lunatech.cmt.Helpers.{exerciseFileHasBeenModified, getFilesToCopyAndDelete, pullTestCode}
+import com.lunatech.cmt.Domain.StudentifiedRepo
 import com.lunatech.cmt.client.Configuration
-import com.lunatech.cmt.client.Domain.{ForceMoveToExercise, StudentifiedRepo}
+import com.lunatech.cmt.client.Domain.ForceMoveToExercise
 import com.lunatech.cmt.client.command.getCurrentExerciseId
 import com.lunatech.cmt.core.validation.Validatable
 import sbt.io.syntax.*
@@ -39,7 +40,6 @@ object NextExercise:
 
   extension (options: NextExercise.Options)
     def execute(configuration: Configuration): Either[CmtError, String] = {
-      import com.lunatech.cmt.client.Domain.ForceMoveToExercise
       val cMTcConfig = new CMTcConfig(options.studentifiedRepo.getOrElse(configuration.currentCourse.value).value)
 
       val currentExerciseId = getCurrentExerciseId(cMTcConfig.bookmarkFile)
