+++ cmtc/src/main/scala/com/lunatech/cmt/client/command/SetCurrentCourse.scala
@@ -1,7 +1,7 @@
 package com.lunatech.cmt.client.command
 
 import caseapp.{AppName, CommandName, ExtraName, HelpMessage, RemainingArgs}
-import com.lunatech.cmt.client.Domain.StudentifiedRepo
+import com.lunatech.cmt.Domain.StudentifiedRepo
 import com.lunatech.cmt.client.cli.ArgParsers.studentifiedRepoArgParser
 import com.lunatech.cmt.client.cli.CmtcCommand
 import com.lunatech.cmt.client.{Configuration, CurrentCourse, findStudentRepoRoot, listExercises}
