+++ cmtc/src/test/scala/com/lunatech/cmt/client/command/SetCurrentCourseSpec.scala
@@ -1,6 +1,6 @@
 package com.lunatech.cmt.client.command
 
-import com.lunatech.cmt.client.Domain.StudentifiedRepo
+import com.lunatech.cmt.Domain.StudentifiedRepo
 import com.lunatech.cmt.client.{Configuration, CurrentCourse}
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.matchers.should.Matchers
