+++ core/src/main/scala/com/lunatech/cmt/Domain.scala
@@ -1,0 +1,10 @@
+package com.lunatech.cmt
+
+import sbt.io.syntax.{File, file}
+
+object Domain {
+
+  final case class StudentifiedRepo(value: File)
+  object StudentifiedRepo:
+    val default: StudentifiedRepo = StudentifiedRepo(file(".").getAbsoluteFile.getParentFile)
+}
