+++ core/src/main/scala/com/lunatech/cmt/client/cli/CmtcCommand.scala
@@ -2,10 +2,10 @@ package com.lunatech.cmt.client.cli
 
 import caseapp.core.help.Help
 import caseapp.core.parser.Parser
-import com.lunatech.cmt.printErrorAndExit
+import cats.syntax.either.*
 import com.lunatech.cmt.client.Configuration
 import com.lunatech.cmt.core.cli.CmtCommand
-import cats.syntax.either.*
+import com.lunatech.cmt.printErrorAndExit
 
 abstract class CmtcCommand[T](using parser: Parser[T], help: Help[T]) extends CmtCommand[T] {
