+++ functional-tests/src/test/scala/com/lunatech/cmt/support/TestSupport.scala
@@ -19,7 +19,7 @@ import com.lunatech.cmt.admin.cli.SharedOptions
 import com.lunatech.cmt.admin.command
 import com.lunatech.cmt.client.{Configuration, CoursesDirectory, CurrentCourse}
 import com.lunatech.cmt.client.Configuration.{CmtHome, GithubApiToken}
-import com.lunatech.cmt.client.Domain.{ExerciseId, ForceMoveToExercise, StudentifiedRepo, TemplatePath}
+import com.lunatech.cmt.Domain.{ExerciseId, ForceMoveToExercise, StudentifiedRepo, TemplatePath}
 import com.lunatech.cmt.client.command.{
   GotoExercise,
   GotoFirstExercise,
